Fixed the config File i had ' insted of ". I also finshed the DailyTempFolderCheckorCreate method.
it should now be vaidating that the file exist at the start up each time.

I need to have something that checks if the App.Config has info in it before i start trying to create the folder.

This will be done latter in the process.
for now im just getting the outline