Some more edit to the doc and the process to check for ticket numbers.
Need to update this process to account for more than just Ticket#
That only works for are CW System but not are Jira System.

Jira has ticket with name like     SSD-231 or SNS-21354
So it will ahve to have multy numer inpout and i will need to have a way to add a list of boards in.
ether as a process or hard coded.
will start with hard coded